// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/get-ot-map-surgery-checklist.dto.ts
export class GetOTMapSurgeryCheckList_DTO {
    SurgeryCheckListId: number = 0;
    SurgeryId: number = 0;
    CheckListId: number = 0;
    DisplaySequence: number = 0;
    IsMandatory: boolean = false;
    CreatedBy: number = 0;
    CreatedOn: string = "";
    ModifiedBy: number = 0;
    ModifiedOn: string = "";
    IsActive: boolean = false;
}