// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/get-ot-surgery.dto.ts
export class GetOTSurgery_DTO {
    SurgeryId: number = 0;
    SurgeryName: string = "";
    SurgeryCode: string = "";
    Description: string = "";
    CreatedOn: string = null;
    CreatedBy: number = null;
    ModifiedOn: string = null;
    ModifiedBy: number = null;
    IsActive: boolean = false;
    IsSystemDefault: boolean = false;
}