// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/temp-map-surgery-checklist.dto.ts
export class TempMapSurgeryCheckList_DTO {
    SurgeryCheckListId: number = 0
    SurgeryId: number = 0
    CheckListId: number = 0;
    IsSelected: boolean = false;
    CheckListName: string = "";
    DisplayName: string = "";
    InputType: string = "";
    IsMandatory: boolean = false;
    IsActive: boolean = false;
    DisplaySequence: number = 0;
    CreatedBy: number = 0;
    CreatedOn: string = "";
    ModifiedBy: number = 0;
    ModifiedOn: string = "";
}